split combined error message component and cleaned up browse page slightly
