+++ web/src/components/forms/pop-form.tsx
@@ -6,6 +6,7 @@ import { ProjectAnnotation } from '../../../types';
 import { useSession } from '../../contexts/session-context';
 import { usePopCreateMutation } from '../../hooks/mutations/usePopCreateMutation';
 import { PepSelector } from './components/pep-selector';
+import { CombinedErrorMessage } from './components/combined-error-message'
 
 interface POPInputs {
   is_private: boolean;
@@ -20,35 +21,6 @@ interface Props {
   onHide: () => void;
   defaultNamespace?: string;
 }
-
-type CombinedErrorMessageProps = {
-  errors: FieldErrors<POPInputs>;
-};
-
-const CombinedErrorMessage = (props: CombinedErrorMessageProps) => {
-  const { errors } = props;
-  const nameError = errors.project_name?.message;
-  const tagError = errors.tag?.message;
-  let msg = null;
-
-  if (nameError == 'empty' && !tagError) {
-    msg = 'Project Name must not be empty.';
-  } else if (nameError == 'invalid' && !tagError) {
-    msg = "Project Name must contain only alphanumeric characters, '-', or '_'.";
-  } else if (nameError == 'empty' && tagError == 'invalid') {
-    msg = "Project Name must not be empty and Tag must contain only alphanumeric characters, '-', or '_'.";
-  } else if (nameError == 'invalid' && tagError == 'invalid') {
-    msg = "Project Name and Tag must contain only alphanumeric characters, '-', or '_'.";
-  } else if (!nameError && tagError == 'invalid') {
-    msg = "Project Tag must contain only alphanumeric characters, '-', or '_'.";
-  }
-
-  if (nameError || tagError) {
-    return <p className="text-danger text-xs pt-1 mb-0">{msg}</p>;
-  }
-
-  return null;
-};
 
 export const PopForm: FC<Props> = ({ onHide, defaultNamespace }) => {
   // get user innfo
@@ -154,7 +126,7 @@ export const PopForm: FC<Props> = ({ onHide, defaultNamespace }) => {
           />
         </div>
       </div>
-      <CombinedErrorMessage errors={errors} />
+      <CombinedErrorMessage errors={errors} formType={'project'} />
       <label className="fw-semibold text-sm mt-2">Description</label>
       <textarea
         id="blank_description"
