+++ web/src/pages/Browse.tsx
@@ -163,7 +163,7 @@ export function Browse() {
 
             <div className="row mt-0">
               <div className="col-12">
-                {topNamespace?.data?.results ? (
+                {topNamespace?.data?.results || isLoading ? (
                   <>
                     <div className="d-flex flex-wrap align-items-center justify-content-between">
                       <div className="fs-6 fw-medium pt-3">Top 10 Starred PEPs</div>
@@ -181,11 +181,11 @@ export function Browse() {
                       Want to see more? Visit the namespace to view remaining projects.
                     </p>
                   </>
-                ) : selectedNamespace ? (
+                ) : (
                   <div className="col-12 mt-4 text-center">
                     <LoadingSpinner />
                   </div>
-                ) : null}
+                )}
               </div>
             </div>
           </>
